+++ NEWS.md
@@ -14,7 +14,8 @@
 
 ### Internals:
 * Between, between, and Within methods: now use ave() instead of tapply().
-* Sum.matrix (non-exported) allows for non-character effect in non-index case.
+* between.martrix and Sum.matrix (the latter is non-exported) allow for
+  non-character effect in non-index case.
 * pmg, pcce: now use the general Between() function of the package (instead of
   "own" between transformation implemented within the respective functions).